display the "stack" in the on_fail callback of is_type
is_type now keeps track on where it currently is inside the value
hierarchy, by storing the object property names and array indices in a
stack. This stack is later used to recreate the "path" from the root
value to the "current" value when on_fail is called, which is included
in the "reason" string.